CSS + Linking
The purpose of this commit was to do the basic CSS for my website, and figure out how to link to the different pages. E.g go to 'Jimmy Donaldson's page from 'MrBeast page. Thirdly, my the images logo was appearing, even when there was no image. Thus I added an if statements.